Organize the code
